pyproject.toml
[tool.poetry]
name = "chinese-hackathon-rag-case"
version = "0.1.0"
description = "This is our solution for a RAG application based on raw HTML files from Chinese news portals. We used Yandex GPT, ChromeDB, Langchain, Spark for solving our case."
authors = ["Ivanov_SV <svyatwork2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
requests = "2.28.1"
opensearch-py = "^2.6.0"
pdf2image = "^1.17.0"
langchain = "^0.2.4"
sentence-transformers = "^3.0.1"
instructorembedding = "^1.0.1"
unstructured = "^0.14.5"
langchain-community = "^0.2.4"
boto3 = "^1.34.126"
fastwarc = "^0.14.7"
langchain-huggingface = "^0.0.3"
html2text = "^2024.2.26"
zhon = "^2.0.2"
nltk = "^3.8.1"
bs4 = "^0.0.2"
python-dotenv = "^1.0.1"
pytelegrambotapi = "^4.19.1"
yandexcloud = "^0.293.0"
PyJWT="2.3.0"
cryptography = "^42.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
